Integrate with org-mode, --load-settings no longer fails on ENOENT
Create an elisp library, gtk-print.el, to allow calling gtk-print from Emacs.
On top of it, build a function to inject an "As PDF and print" option into
org-mode's latex exporter.

If the file specified by --load-settings does not exist, no longer quit. This is
because --load-settings and --save-settings are often used in conjunction, and
the load-settings file could have not been created yet, for example if gtk-print
is called the first time or if the user deleted that file.