Print current buffer using wkhtmltopdf; async interface
Add a new command, `gtk-print-buffer-wkhtmltopdf', which allows the user to
print the current buffer using htmlize + wkhtmltopdf.

Add a new function, `gtk-print-file-async', which can be used to print files
without blocking the GUI thread. Refactor the common command-building code into
a new function, `gtk-print--file-shell-command', to facilitate this.

Add a new password query method, 'getpass-cli', which always prompts with
"Password: ", regardless of language and locale settings. Update the German
translation to address this.