Specify fallback for --settings-file with --load-settings instead
Previously, --load-settings would override --settings-file's file for loading.
This is redundant, as the same effect could already be achieved by specifying
--load-settings with --save-settings. If the two options are specified together,
--load-settings's file is used as a fallback, if the file specified by
--settings-file doesn't exist. This way, a default set of print options can be
configured for the first run. Adjust the REAMDE. If you want to load the
settings from a file that may not exist, specify it as the target for
--settings-file and change the output to some other file with --save-settings.

Adjust .clang-format to force binary operators other than assignments to the
right.